Rename and move create… and translateArticle.js files
This patch does nothing but renaming/moving files. The only purpose of
this patch is so that we can rebase Ib0c8694 on top of this and have
actual diffs there, instead of deleted and added files.

Note that this patch merges two directories, but leaves the two
ResourceLoader modules intact. This is meant to be temporary.

Bug: T124036
Change-Id: I4f5db3bb63cbe1ffadf8d1789d08b26ed0921ed3