Prevent overlap of sidebar column with everything else
The .articleplaceholder-sidebar is floated to the left. This is what we
want. However, if the .articleplaceholder-identifierlist does not
contain many items then statement-groups can appear under it. This
creates a slightly off-topic look to the page. This commit should
prevent this by setting the statement-group list to `width: auto` and
`overflow: hidden`, as per https://stackoverflow.com/a/5195902/5013691.

Note that this must happen only when the viewport is at least 912px
wide, or the statement-groups will be partially cut off due to their
fixed size.

Bug: T149193
Change-Id: I295fccdf476e739eadb4ff29797691ba732fda6b