Properly catch MalformedTitleException in AboutTopicRenderer
I'm opening a new control flow path here:

When the entity's label forms a valid page name, the code checks if the
user is allowed to create this page, and only displays the "create
article" button if this is the case. The user can change the page name
in the dialog.

Note this is intentional. When the original label got blocked on the
local wiki, users should not be invited to create the page again, not
even with an other title.

The new code path in this patch leaves the page title field empty in case
the label does not form a valid page title. In this case user rights are
intentionally not checked in advance.

Bug: T181562
Change-Id: Ic7cf6244d980862ca091e5d3af40569f5bd5f188