Remove grunt-contrib-qunit test runner
ArticlePlaceholder should use the MediaWiki Qunit runner. An attempt to
port the runner is https://gerrit.wikimedia.org/r/#/c/444936/ but it is
not as simple as envisionned.

It is the last repository on Wikimedia CI to use legacy jobs, and they
have to be removed.  To unblock the migration:

* drop the grunt-contrib-qunit package and related dependencies.
* leave a note in tests/qunit/README indicating tests are not being run

The port can be achieved via the other change.

Bug: T180171
Bug: T183512
Change-Id: Ia1c6989757d98e46cb4695ea20bd07bf39e7da01