Fix broken sinon.stub use in createArticleTranslation.test.js
Looks like a combination of a bad git-merge conflict resolution,
and accidental assignment where stub() parameters should have
been used.

A plain assignment is not known to Sinon and cannot be restored.
We can't use Sinon here because mw.cx itself may be undefined.

This test is written in a mock-everything style, as a combination
between integration and unit testing. I'll try to follow that
style for now by defining the object entirely (as before),
but also restoring it, so that it won't break other unit
tests in this extension, or another.

Bug: T219663
Change-Id: I390ee61dc390f2efdaf45cc6832ce6c4282fb591