Replace custom Wikibase with upstream MediaWiki CodeSniffer
As argued in Ic60d497: This is not a codebase the Wikidata team(s) are
currently actively working on. Having the custom Wikibase CodeSniffer
here does not provide a benefit any more, as far as I can tell, but
does make it harder for others to contribute and do mass-updates like
updating the CodeSniffer rule set.

Change-Id: I2aa9dde3d5d8647180151589d086c933044b8734