Temporarily skip test
The WikibaseClient Store will move to the service container, and I don’t
think it’s feasible to keep overrideStore() in that case, so that method
will have to be removed. Skip this test so that Wikibase can remove the
method without breaking the test, and then this test will be updated in
a follow-up.

Bug: T277482
Change-Id: Ife404b9daab1b503426c7220019b245323b31c00