tests: Fix `assert.expect(N)` and enable wikimedia/qunit eslint preset
The wikimedia/qunit preset (and the eslint-plugin-qunit package) both
enforce the "require-expect=never-except-zero" rule by default, but
this repo didn't enable that preset yet.

See also https://qunitjs.com/api/assert/expect/.
Basically, if you don't have anti-patterns like nesting assertions
in externally controlled branches or callbacks, the only way for an
assertion to not be reached, is for the test to already be failing.
The these anti-patterns are fortunately rare and unneeded. Especially
with assert.rejects(), assert.async(), and async-await support
nowadays this is simply unneeded busywork. As demonstrated by:

This is one of only 4 WMF-deployed repos that still use it this way:
https://codesearch.wmcloud.org/deployed/?q=assert.expect%5C%28%5Cs%2A%5B%5E0%20%5D&files=js%24

Change-Id: Ib13f285b301d0769d8ec45707e7eaf9e81137e19