build: Re-enable eslint-preset-wikimedia rules
The vast majority of the eslint rules got lost for this extension in
recent years, because the eslint-preset-wikimedia package has spit up
the settings into smaller presets, with much less in the "wikimedia"
base preset than there used to be.

* Use client+jquery+jquery for most code, and the all-in-one
  "mediawiki-qunit" preset for the tests.
* Disable any new rules that pass or don't apply.
* Remove redundant settings that are now inherited.

The one thing that I fixed within this patch is the use of obsolete
globals (jQuery, mediaWiki) since the preferred form "$" and "mw"
are not only equally available as globals, they are also explicitly
injected into the module closure by ResourceLoader and thus actually
(marginally) faster to reference. This matches the way other extensions
and core have written JavaScript files since ~2013.

Change-Id: Id2ee17160e5d592ca82f2fa0ddf0e9dfc803137f