tests: Use this.sandbox instead of sinon.sandbox.create()
Extension don't need to call this method, as testrunner.js in MediaWiki
already creates a new sandbox for every test. Only a handful of
extensions do this, presumably without realizing [1].

This is motivated by a breaking change to sinon.sandbox.create()
which blocks the Sinon 18 upgrade, and can by avoiding by using the
built-in Sinon sandbox.

[1]: https://codesearch.wmcloud.org/deployed/?q=sinon.sandbox.create&excludeFiles=node_mod%7Cnode-qunit

Bug: T389450
Change-Id: I2f15a196fa9e7e431077b5a57b14c729d6714907