ItemNotabilityFilter: Remove use of SessionConsistentConnectionManager
This was introduced in 2016 with I9b674b8f2b, which converted an API
query to a DB query, using Wikibase ConsistentReadConnectionManager,
a predecessor to SessionConsistentConnectionManager in MediaWiki core.

At this time, LoadBalancer->getConnection still defaulted to an
eager-connected manually-released "Database" object, which thus made
it important to call that as late as possible, which have you two
bad options:

1. Inject LBFactory and the caller has to be careful to call both
   getMainLB and getConnection which is both verbose (two calls)
   and fragile (pass the right DB domain to both, or else it
   works locally but blows up in production).
2. Inject LoadBalancer and $domain, with a similar footgun.
3. Inject DBConnRef (doable but rare in practice, not sure why).

Fast-forward to 2025, and getConnection now defaults to an auto-reused
DBConnRef that is always lazy-connecting (T255493), and we also
have the narrower ConnectionProvider interface to LBFactory (T330641).

This code did not seem to use SessionConsistentConnectionManager for
its primary vs replica switching logic.

Change-Id: Ia279f361e12ec9d45b0602587350ea5bc38082e8